Add CILogon client and configuration changes
This commit creates new CILogon client entities and updates the application configuration for dealing with CILogon operations. This includes new CILogon client, request, response, and institution classes, as well as modifications to the application's YAML file to support these changes. These additions and modifications allow the application to interact with the CILogon API more seamlessly.

closes #36